Create: custom Hook and move related code into it
